# target/generated-sources/openapi/src/main/java/com/gila/ecommerce/api/DefaultApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.gila.ecommerce.api;

import com.gila.ecommerce.dto.CartDto;
import com.gila.ecommerce.dto.CartItemRequestDto;
import com.gila.ecommerce.dto.ImportProducts202Response;
import com.gila.ecommerce.dto.LoginRequestDto;
import com.gila.ecommerce.dto.LoginResponseDto;
import com.gila.ecommerce.dto.OrderDto;
import com.gila.ecommerce.dto.ProductDto;
import com.gila.ecommerce.dto.ProductImportStatusDto;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-10T21:50:03.681471300-06:00[GMT-06:00]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Default", description = "the Default API")
public interface DefaultApi {

    /**
     * POST /orders/checkout : Process checkout for the current cart (Customer only)
     *
     * @return Order checked out and payment authorized successfully (status code 200)
     *         or Empty cart or stock validation failed (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (requires customer role) (status code 403)
     */
    @Operation(
        operationId = "checkout",
        summary = "Process checkout for the current cart (Customer only)",
        responses = {
            @ApiResponse(responseCode = "200", description = "Order checked out and payment authorized successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrderDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Empty cart or stock validation failed"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires customer role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/orders/checkout",
        produces = { "application/json" }
    )
    
    ResponseEntity<OrderDto> checkout(
        
    );


    /**
     * DELETE /cart : Clear all items from the cart (Customer only)
     *
     * @return Cart cleared successfully (status code 204)
     *         or Forbidden (requires customer role) (status code 403)
     */
    @Operation(
        operationId = "clearCart",
        summary = "Clear all items from the cart (Customer only)",
        responses = {
            @ApiResponse(responseCode = "204", description = "Cart cleared successfully"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires customer role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/cart"
    )
    
    ResponseEntity<Void> clearCart(
        
    );


    /**
     * DELETE /orders/clear : Truncate database orders and reset system inventory defaults (Admin only)
     *
     * @return Orders database reset completed successfully (status code 204)
     *         or Forbidden (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "clearOrders",
        summary = "Truncate database orders and reset system inventory defaults (Admin only)",
        responses = {
            @ApiResponse(responseCode = "204", description = "Orders database reset completed successfully"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/orders/clear"
    )
    
    ResponseEntity<Void> clearOrders(
        
    );


    /**
     * POST /products : Create a new product (Admin only)
     *
     * @param productDto  (required)
     * @return Product successfully created (status code 201)
     *         or Invalid product details supplied (status code 400)
     *         or Forbidden operation (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "createProduct",
        summary = "Create a new product (Admin only)",
        responses = {
            @ApiResponse(responseCode = "201", description = "Product successfully created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid product details supplied"),
            @ApiResponse(responseCode = "403", description = "Forbidden operation (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/products",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProductDto> createProduct(
        @Parameter(name = "ProductDto", description = "", required = true) @Valid @RequestBody ProductDto productDto
    );


    /**
     * DELETE /products/{id} : Delete a product (Admin only)
     *
     * @param id  (required)
     * @return Product deleted successfully (status code 204)
     *         or Product not found (status code 404)
     *         or Forbidden (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "deleteProduct",
        summary = "Delete a product (Admin only)",
        responses = {
            @ApiResponse(responseCode = "204", description = "Product deleted successfully"),
            @ApiResponse(responseCode = "404", description = "Product not found"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/products/{id}"
    )
    
    ResponseEntity<Void> deleteProduct(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /cart : Get the current customer&#39;s shopping cart (Customer only)
     *
     * @return Shopping cart retrieved successfully (status code 200)
     *         or Unauthorized JWT session (status code 401)
     *         or Forbidden (requires customer role) (status code 403)
     */
    @Operation(
        operationId = "getCart",
        summary = "Get the current customer's shopping cart (Customer only)",
        responses = {
            @ApiResponse(responseCode = "200", description = "Shopping cart retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartDto.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized JWT session"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires customer role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/cart",
        produces = { "application/json" }
    )
    
    ResponseEntity<CartDto> getCart(
        
    );


    /**
     * GET /products/import/status/{taskId} : Check status and validation reports of a CSV import task (Admin only)
     *
     * @param taskId  (required)
     * @return Task import status report retrieved (status code 200)
     *         or Import task not found (status code 404)
     *         or Forbidden (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "getImportStatus",
        summary = "Check status and validation reports of a CSV import task (Admin only)",
        responses = {
            @ApiResponse(responseCode = "200", description = "Task import status report retrieved", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductImportStatusDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Import task not found"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products/import/status/{taskId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<ProductImportStatusDto> getImportStatus(
        @Parameter(name = "taskId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("taskId") UUID taskId
    );


    /**
     * GET /products/{id} : Get details of a single product
     *
     * @param id  (required)
     * @return Product details retrieved (status code 200)
     *         or Product not found (status code 404)
     */
    @Operation(
        operationId = "getProductById",
        summary = "Get details of a single product",
        responses = {
            @ApiResponse(responseCode = "200", description = "Product details retrieved", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<ProductDto> getProductById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /products : Retrieve products with optional filter/search
     *
     * @param query Search term matching product name or description (optional)
     * @param category Filter products by category (optional)
     * @param page Page number (0-indexed) (optional, default to 0)
     * @param size Page size limit (optional, default to 10)
     * @return List of products matching criteria (status code 200)
     */
    @Operation(
        operationId = "getProducts",
        summary = "Retrieve products with optional filter/search",
        responses = {
            @ApiResponse(responseCode = "200", description = "List of products matching criteria", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ProductDto.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<ProductDto>> getProducts(
        @Parameter(name = "query", description = "Search term matching product name or description", in = ParameterIn.QUERY) @Valid @RequestParam(value = "query", required = false) String query,
        @Parameter(name = "category", description = "Filter products by category", in = ParameterIn.QUERY) @Valid @RequestParam(value = "category", required = false) String category,
        @Parameter(name = "page", description = "Page number (0-indexed)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "Page size limit", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "10") Integer size
    );


    /**
     * POST /products/import : Asynchronously import products from a CSV file (Admin only)
     *
     * @param file The CSV file to upload containing product records (optional)
     * @return CSV import task accepted and queued for Kafka processing (status code 202)
     *         or Invalid file format or missing CSV headers (status code 400)
     *         or Forbidden (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "importProducts",
        summary = "Asynchronously import products from a CSV file (Admin only)",
        responses = {
            @ApiResponse(responseCode = "202", description = "CSV import task accepted and queued for Kafka processing", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ImportProducts202Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid file format or missing CSV headers"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/products/import",
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    
    ResponseEntity<ImportProducts202Response> importProducts(
        @Parameter(name = "file", description = "The CSV file to upload containing product records") @RequestPart(value = "file", required = false) MultipartFile file
    );


    /**
     * POST /auth/login : Authenticate user and return JWT
     *
     * @param loginRequestDto  (required)
     * @return Successful authentication (status code 200)
     *         or Unauthorized credentials (status code 401)
     */
    @Operation(
        operationId = "login",
        summary = "Authenticate user and return JWT",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful authentication", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LoginResponseDto.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized credentials")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<LoginResponseDto> login(
        @Parameter(name = "LoginRequestDto", description = "", required = true) @Valid @RequestBody LoginRequestDto loginRequestDto
    );


    /**
     * DELETE /cart/items/{productId} : Remove a specific item from the cart (Customer only)
     *
     * @param productId  (required)
     * @return Item removed and updated cart returned (status code 200)
     *         or Product not found in cart (status code 404)
     *         or Forbidden (requires customer role) (status code 403)
     */
    @Operation(
        operationId = "removeCartItem",
        summary = "Remove a specific item from the cart (Customer only)",
        responses = {
            @ApiResponse(responseCode = "200", description = "Item removed and updated cart returned", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product not found in cart"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires customer role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/cart/items/{productId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<CartDto> removeCartItem(
        @Parameter(name = "productId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("productId") UUID productId
    );


    /**
     * POST /cart : Add or update item quantity in the cart (Customer only)
     *
     * @param cartItemRequestDto  (required)
     * @return Shopping cart updated successfully (status code 200)
     *         or Invalid request payload or quantity limits exceeded (status code 400)
     *         or Forbidden (requires customer role) (status code 403)
     */
    @Operation(
        operationId = "updateCartItem",
        summary = "Add or update item quantity in the cart (Customer only)",
        responses = {
            @ApiResponse(responseCode = "200", description = "Shopping cart updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payload or quantity limits exceeded"),
            @ApiResponse(responseCode = "403", description = "Forbidden (requires customer role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/cart",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<CartDto> updateCartItem(
        @Parameter(name = "CartItemRequestDto", description = "", required = true) @Valid @RequestBody CartItemRequestDto cartItemRequestDto
    );


    /**
     * PUT /products/{id} : Update an existing product (Admin only)
     *
     * @param id  (required)
     * @param productDto  (required)
     * @return Product updated successfully (status code 200)
     *         or Product not found (status code 404)
     *         or Forbidden operation (requires admin role) (status code 403)
     */
    @Operation(
        operationId = "updateProduct",
        summary = "Update an existing product (Admin only)",
        responses = {
            @ApiResponse(responseCode = "200", description = "Product updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product not found"),
            @ApiResponse(responseCode = "403", description = "Forbidden operation (requires admin role)")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/products/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProductDto> updateProduct(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "ProductDto", description = "", required = true) @Valid @RequestBody ProductDto productDto
    );

}